+++ scripts/Plasma Noise.py
@@ -101,7 +101,7 @@ class Script(scripts.Script):
         p.extra_generation_params["Blue Min"] = blu_min
         p.extra_generation_params["Blue Max"] = blu_max
         p.initial_noise_multiplier = noise_mult
-        p.denoising_strength = int(denoising)
+        p.denoising_strength = float(denoising)
 
         w = p.width
         h = p.height
